Generate trustee keys and update election config via election API (#15)
Prior to this change, the app was not connected to the election API and faked the configuration and trustee data. This change actually wires up the API call to generate the individual trustee keys, and to update the election configuration, including the joint public key, and total number of ballot selections.