Use election API to record ballots
Prior to this change, ballots were stubbed. This change wires up the
integration with the election API so that they can be recorded for
tallying.

Note: the export path has been altered. It seems the API does not
support the use of '~' as the home shortcut. It has been changed to the
globals path delimiter to make supporting different OSes easier.