✨Implement Print of Tally Result File
The Tally Result can now be printed as displayed by the React component to ensure the final result and the printout are similar. This uses the React to Print functionality to make this flexible to quick changes later. 

The linting ignores are necessary due to the react-to-print function. The package does not support functional components and ref requires the use of a null instead of undefined.