Write election configuration to encryptor drives (#49)
* Provide usb presence and mounted statuses separately

Prior to this change, we were setting the usb status based on whether or
not it was currently mounted. Since usb drives auto-mount, this change
updates the status logic to split mounted and availability, so that we
can detect when users remove a device from the machine.

* Write election config to encryptor drives

* 🔧Usb Mounting Procedure and Mock 

Update the mock to allow for handling mocking usb mounting. 
Usb Manager now allows the mounting of the drives.

* fix merge error

Co-authored-by: Keith Fung <keithrfung@users.noreply.github.com>
Co-authored-by: Matt Wilhelm <AddressXception@users.noreply.github.com>